feat: trans to rapidTable code style
