+++ wired_table_rec/__init__.py
@@ -2,6 +2,6 @@
 # @Author: SWHL
 # @Contact: liekkaskono@163.com
 from .main import WiredTableRecognition
-from .utils_table_recover import vis_table
+from wired_table_rec.utils.utils_table_recover import vis_table
 
 __all__ = ["WiredTableRecognition", "vis_table"]
